// src/app/models/PojoContratista.ts
export class PojoContratista{
  
    userName: string;
    cedula: string;
    rol:string;
	passWord: string;
	nam_AREA: string;
	name_EPS: string;
	pension: string;
	nam_ARL: string;
	name_BANCO: string;
	tipo_CUENTA: string;
	name_SUPER: string;
	id_RIESGO: string;
	id_PROYECTO: string;
	n_CONTRATO: string;
	nombres: string;
	telefono: string;
	direccion: string;
	n_CUENTA_BANCO: string;
	cedula_SUPERVISOR: string;
	ciudad_cedula_supervisor: string;
	cargo: string;
	detalle_POR:string;
	firma_Contratista: string;
	
}