// src/app/models/PojoContrato.ts
export class PojoContrato{

      userName: string;
	  cedula: string;
	  n_CONTRATO: string;
	  fecha_INICIO_CONTRATO: string;
	  fecha_FINAL_CONTRATO: string;
	  valor_TOTAL_CONTRATO: string;     //moneda tipo variable ¿?
	  valor_CONTRATO: string;			//moneda tipo variable ¿?
	  detalle: string;
	  valor_ADICION: string;			//moneda tipo variable ¿?
	  detalle_PRORROGA: string;
	  id_SIPSE: string;
	  detalle_SUPENCION: string;
	  objeto_CONTRATO: string;
	  observaciones_CONTRATO: string;
	  tiempo_MESES: string;
	  tiempo_DIAS: string;
	  duracion_CONTRATO: string;
	  detalle_POR: string;

    

}