// src/app/models/PojoObligaciones2.ts
export class PojoObligaciones2{
  
    userName: string;
    nombres: string;
    n_CONTRATO: String;
    idsipse: string;
    obligacion_1: string;
    obligacion_2: string;
    obligacion_3: string;
    obligacion_4: string;
    obligacion_5: string;
    obligacion_6: string;
    obligacion_7: string;
    obligacion_8: string;
    obligacion_9: string;
    obligacion_10: string;
    obligacion_11: string;
    obligacion_12: string;
    obligacion_13: string;
    obligacion_14: string;
    obligacion_15: string;
    obligacion_16: string;
    obligacion_17: string;
    obligacion_18: string;
    obligacion_19: string;
    
}